попытка в автотесты №3
