--- src/tests/notTest.php
@@ -1,1 +1,0 @@
-<?php
