--- src/tests/rus.php
@@ -1,1 +1,0 @@
-<?php
